Cosmos: GlobalCommittedLSN parsing (#389)
Porting - https://github.com/Azure/azure-sdk-for-java/pull/21937

+++ direct-impl/src/main/java/com/microsoft/azure/cosmosdb/internal/directconnectivity/StoreReader.java
@@ -787,7 +787,7 @@ public class StoreReader {
 
                 headerValue = documentClientException.getResponseHeaders().get(WFConstants.BackendHeaders.GLOBAL_COMMITTED_LSN);
                 if (!Strings.isNullOrEmpty(headerValue)) {
-                    globalCommittedLSN = Integer.parseInt(headerValue);
+                    globalCommittedLSN = Long.parseLong(headerValue);
                 }
 
                 long lsn = -1;
